requirements.txt
ariadne ==0.17.1
python-dotenv== 0.21.0
mongoengine==0.25.0
